Librarian and LibrarianGender Controller

+++ src/main/java/za/ac/cput/service/impl/LibrarianService.java
@@ -4,7 +4,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import za.ac.cput.domain.Librarian;
 import za.ac.cput.repository.LibrarianRepository;
+import za.ac.cput.repository.LibrarianRepository;
 
+import java.util.ArrayList;
 import java.util.List;
 @Service
 public class LibrarianService implements LibrarianIservice {
@@ -31,13 +33,45 @@ private final LibrarianRepository librarianRepository;
 
     @Override
     public void delete(String id) {
-    this.librarianRepository.deleteById(id);
+    if (this.librarianRepository.existsById(id)){
+        this.librarianRepository.deleteById(id);
+        System.out.println("Deleted successfully");
+
+
+
+    }else{
+        System.out.println("Could not find Id specified");
+
+    }
+
+
 
     }
 
     @Override
     public List<Librarian> getAll() {
         return this.librarianRepository.findAll();
+    }
+
+
+    public List<Librarian>  findLibrarianByPosition(String position) {
+        List<Librarian> subset= this.librarianRepository.findLibrarianByPosition(position);
+        List<Librarian> librarianList = new ArrayList<>();
+        for (Librarian l:subset){
+            librarianList.add(l);
+        }
+        return librarianList;
+    }
+    public List<Librarian> findLibrarianByName(String name){
+
+        List<Librarian> subset= this.librarianRepository.findLibrarianByName(name);
+        List<Librarian> librarianList = new ArrayList<>();
+        for (Librarian l:subset){
+            librarianList.add(l);
+        }
+        return librarianList;
+
+
     }
 }
